requirements.txt
python-telegram-bot>=20.0
google-api-python-client>=2.0
google-auth-oauthlib>=1.0
google-auth-httplib2>=0.2.0
gspread>=6.0
groq>=0.4.0
python-dotenv>=1.0
